refactored and cleaned up save-dataset
